+++ eslint.config.ts
@@ -2,6 +2,7 @@ import eslint from "@eslint/js";
 import tseslint from "@typescript-eslint/eslint-plugin";
 import tsparser from "@typescript-eslint/parser";
 import prettierConfig from "eslint-config-prettier";
+import globals from "globals";
 
 export default [
   eslint.configs.recommended,
@@ -14,7 +15,12 @@ export default [
         sourceType: "module",
       },
       globals: {
-        console: "readonly",
+        ...globals.browser,
+        ...globals.node,
+        // Additional Web API types that might not be in the standard globals
+        BodyInit: "readonly",
+        RequestInit: "readonly",
+        HeadersInit: "readonly",
       },
     },
     plugins: {